% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{penguins}
\alias{penguins}
\title{Palmer penguins dataset}
\format{
A data frame with 344 rows and 8 columns:
\describe{
\item{species}{Penguin species: Adelie, Chinstrap, or Gentoo.}
\item{island}{Island where the penguin was observed.}
\item{bill_len}{Bill length in millimetres.}
\item{bill_dep}{Bill depth in millimetres.}
\item{flipper_len}{Flipper length in millimetres.}
\item{body_mass}{Body mass in grams.}
\item{sex}{Penguin sex.}
\item{year}{Year of observation.}
}
}
\source{
Horst AM, Hill AP, Gorman KB (2020).
Palmer Archipelago penguin data.
Available at \url{https://allisonhorst.github.io/palmerpenguins/}.
}
\usage{
data(penguins)
}
\description{
Measurements for penguins observed in the Palmer Archipelago, Antarctica.
}
\keyword{datasets}
