% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestPack-package.R
\docType{package}
\name{TestPack-package}
\alias{TestPack}
\alias{TestPack-package}
\title{TestPack: Package to test package development in R}
\description{
Test the R package development workflow.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/alre1394/mlpack}
  \item Report bugs at \url{https://github.com/alre1394/mlpack/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

Authors:
\itemize{
  \item First Last \email{first.last@example.com}
}

}
\keyword{internal}
